+++ CHANGELOG.md
@@ -63,6 +63,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Raise an error for an empty key in metadata - #6176 by @IKarbowiak
 - Add attributes to product error - #6181 by @IKarbowiak
 - Allow to add product variant with 0 price to draft order - #6189 by @IKarbowiak
+- Fix deleting product when default variant is deleted - #6186 by @IKarbowiak
 
 ## 2.10.2